Re-engineer WebRequester
This is a big commit, that cleans up a few things done wrong in the past. First
of, ContentType.Negotiated is now removed, instead this is configured in the
Advanced configuration section.

Secondly, all the akward classes for sending request (lite the RequestSpec,
RequestTool etc) is replaced by a WebRequester, that does all the things, but
in much fewer lines of code.

Lastly, the concept of 'RequestResender' is removed -- the only resender that
exists today is the UnsupportedMediaType, and it is nicely handled in the new
WebRequesterakward classes for sending request (lite the RequestSpec,
RequestTool etc) is replaced by a WebRequester, that does all the things, but
in much fewer lines of code.

Lastly, the concept of 'RequestResender' is removed -- the only resender that
exists today is the UnsupportedMediaType, and it is nicely handled in the new
WebRequester.